requirements.txt
transformers==4.48.3
pillow==11.0.0